# areas/backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "sean",
    "version": "1.0.0",
    "title": "Cloud service",
    "contact": {
      "email": "grigorevmp@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/department": {
      "get": {
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page for department",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page for department",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/department_info"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/department_info"
              }
            }
          }
        },
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Already exists"
          },
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/department_info"
              }
            }
          }
        },
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Already exists"
          },
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/department/users": {
      "get": {
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name for department",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Failed. Not found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/department_with_users"
                }
              }
            }
          }
        }
      },
      "post": {
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name for department",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          }
        },
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Failed. Not found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/department_with_users"
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Name for department",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users"
              }
            }
          }
        },
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Failed. Not found"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/department_with_users"
                }
              }
            }
          }
        }
      }
    },
    "/all_workspaces": {
      "get": {
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page for department",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page for department",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "deleted",
            "in": "query",
            "description": "Show deleted workspaces",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "Get all workspaces and its owners",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/workspace"
                }
              }
            }
          }
        }
      }
    },
    "/workspace/{space_id}": {
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/workspace_update"
              },
              "example": {
                "new_status": "1",
                "new_owner": "8e6808f7-d29c-422f-bab1-2a8b9b11d018"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "description": "ID of workspace to update",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "Update workspace status or owner",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "404": {
            "description": "Failed. Not found error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/workspace"
                }
              }
            }
          }
        }
      },
      "delete": {
        "parameters": [
          {
            "name": "space_id",
            "in": "path",
            "description": "ID of workspace to delete",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "Delete workspace",
        "responses": {
          "400": {
            "description": "Failed. Invalid workspace ID"
          },
          "404": {
            "description": "Failed. Not found error"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/workspace"
                }
              }
            }
          }
        }
      }
    },
    "/user": {
      "get": {
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page for users",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page for users",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "tags": [
          "AdminRequest"
        ],
        "summary": "to do",
        "responses": {
          "400": {
            "description": "Failed. Bad put data"
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/user_info"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "workspace": {
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        }
      },
...
      "department_info": {
        "properties": {
          "department_name": {
            "type": "string"
          }
        }
      },
      "user_info": {
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "department_with_users": {
        "properties": {
          "department_name": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/user_info"
            }
          }
        }
      },
      "users": {
        "properties": {
          "users": {
            "type": "array",
            "items": {
               "$ref": "#/components/schemas/user_id"
            }
          }
        }
      },
      "user_id": {
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "workspace_update": {
        "properties": {
          "new_status": {
            "type": "string"
          },
          "new_owner": {
            "type": "string"
          }
        }
      }
    }
  }
}